implement_correct_result
