+++ src/components/BookList.1.jsx
@@ -8,85 +8,103 @@ import {
   collection,
   deleteDoc,
   doc,
+  getDocs,
   onSnapshot,
   orderBy,
   query,
 } from "firebase/firestore";
 import { database } from "../firebase";
 import { useState } from "react";
+import useFirestore from "../hooks/useFirestore";
+import { useContext } from "react";
+import { AuthContext } from "../Contexts/AuthContext";
+import { useNavigate } from "react-router-dom";
 
 export const BookList = () => {
   let location = useLocation();
   let params = new URLSearchParams(location.search);
-
   let search = params.get("search");
-  // let url = `http://localhost:2801/books/${search ? `?q=${search}` : ""}`;
-  // const { data: books, loading, error } = useFetch(url);
-  const [books, setBooks] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState("");
 
-  useEffect(function () {
-    setLoading(true);
-    let books = [];
-    let ref = collection(database, "books");
-    let qry = query(ref, orderBy("date", "desc"));
-    // getDocs(qry).then((data) => {
-    //   if (data.empty) {
-    //     setError("No result found");
-    //     setLoading(false);
-    //   }
-    //   if (data.empty == false) {
-    //     data.forEach((d) => {
-    //       books.push({ id: d.id, ...d.data() });
-    //     });
-    //     setBooks(books);
-    //     setLoading(false);
-    //     setError("");
-    //   }
-    // });
-    // real time transition firebase
-    onSnapshot(qry, (data) => {
-      if (data.empty) {
-        setError("No result found");
-        setLoading(false);
-      }
-      if (data.empty == false) {
-        data.forEach((d) => {
-          books.push({ id: d.id, ...d.data() });
-        });
-        setBooks(books);
-        setLoading(false);
-        setError("");
-      }
-    });
-  }, []);
+  let { getCollection, deleteDocument } = useFirestore();
+  let { user } = useContext(AuthContext);
+  console.log(user);
+  let {
+    error,
+    data: books,
+    loading,
+  } = getCollection("books", ["uid", "==", user.uid]);
 
-  const deleteBook = async (e, id) => {
+  let navigate = useNavigate();
+
+  let deleteBook = async (e, id) => {
     e.preventDefault();
-    let ref = doc(database, "books", id);
-    await deleteDoc(ref);
-    setBooks((prev) => prev.filter((p) => p.id != id));
+    await deleteDocument("books", id);
   };
+  // setBooks((prev) => prev.filter((d) => d.id !== id));
+
   let { isDark } = useTheme();
+
+  if (error) {
+    return <p className="text-center text-2xl text-gray-500">{error}</p>;
+  }
+
+  // useEffect(function () {
+  //   setLoading(true);
+  //   let books = [];
+  //   let ref = collection(database, "books");
+  //   let qry = query(ref, orderBy("date", "desc"));
+  //   // getDocs(qry).then((data) => {
+  //   //   if (data.empty) {
+  //   //     setError("No result found");
+  //   //     setLoading(false);
+  //   //   }
+  //   //   if (data.empty == false) {
+  //   //     data.forEach((d) => {
+  //   //       books.push({ id: d.id, ...d.data() });
+  //   //     });
+  //   //     setBooks(books);
+  //   //     setLoading(false);
+  //   //     setError("");
+  //   //   }
+  //   // });
+  //   // real time transition firebase
+  //   onSnapshot(qry, (data) => {
+  //     if (data.empty) {
+  //       setError("No result found");
+  //       setLoading(false);
+  //     }
+  //     if (data.empty == false) {
+  //       data.forEach((d) => {
+  //         books.push({ id: d.id, ...d.data() });
+  //       });
+  //       setBooks(books);
+  //       setLoading(false);
+  //       setError("");
+  //     }
+  //   });
+  // }, []);
+
+  // const deleteBook = async (e, id) => {
+  //   e.preventDefault();
+  //   let ref = doc(database, "books", id);
+  //   await deleteDoc(ref);
+  //   setBooks((prev) => prev.filter((p) => p.id != id));
+  // };
+  // let { isDark } = useTheme();
 
   return (
     <>
       <div className="grid grid-cols-2 md:grid-cols-4 gap-3 my-2 ml-9">
-        {error && <div>{error}</div>}
         {loading && <div className="">Loading ...</div>}
         {!!books &&
           books.map((data) => (
             <Link to={`http://localhost:5173/books/${data.id}`} key={data.id}>
               <div
                 className={`p-1 border-1 border w-[180px] ${
-                  isDark
-                    ? "bg-dcard border-primary text-white"
-                    : "bg-transparent"
+                  isDark ? "bg-dcard border-primary text-white" : ""
                 }`}
               >
-                {/* <img src={user} /> */}
-                <img src={data.image} alt="" className="w-full h-[220px]" />
+                <img src={data.cover} alt="" className="w-full h-[220px]" />
                 <div className="space-y-2 text-center">
                   <div className="mt-2 text-md">{data.title} </div>
                   <div className="flex  m-1 justify-between items-center">
@@ -103,6 +121,7 @@ export const BookList = () => {
                       <Link to={`edit/${data.id}`}>
                         <img src={edit} alt="" />
                       </Link>
+
                       <img
                         src={trash}
                         alt=""
